у меня поехала крыша)

+++ src/main/java/hiber/service/UserService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> getUsers();
-    User getUserById(int id);
+    User getUserById(long id);
     void updateUser(User user);
     void deleteUser(long id);
